extract determine_new_blocks into a separate utility
